requirements.txt
python-multipart
opencv-python
tensorflow
numpy
pillow
pydantic
fastapi